unify the generation location of Mock in one place
