+++ test/mocks/mock_weather_data_source.dart
@@ -1,0 +1,8 @@
+// ignore_for_file: unused_import
+
+import 'package:flutter_training/data/data_source/weather_data_source.dart';
+import 'package:mockito/annotations.dart';
+import 'package:yumemi_weather/yumemi_weather.dart';
+
+@GenerateNiceMocks([MockSpec<WeatherDataSource>()])
+import 'mock_weather_data_source.mocks.dart';
