+++ test/mocks/mock_weather_data_source.mocks.dart
@@ -1,14 +1,13 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in flutter_training/test/unit_test/data/use_case/fetch_weather_use_case_test.dart.
+// in flutter_training/test/mocks/mock_weather_data_source.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:flutter_training/data/models/app_api_result.dart' as _i2;
-import 'package:flutter_training/data/models/weather/weather.dart' as _i4;
-import 'package:flutter_training/data/models/weather/weather_request.dart'
-    as _i5;
-import 'package:flutter_training/data/repository/weather_repository.dart'
+import 'package:flutter_training/data/data_source/weather_data_source.dart'
     as _i3;
+import 'package:flutter_training/data/models/weather/weather.dart' as _i2;
+import 'package:flutter_training/data/models/weather/weather_request.dart'
+    as _i4;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -22,9 +21,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeAppApiResult_0<T> extends _i1.SmartFake
-    implements _i2.AppApiResult<T> {
-  _FakeAppApiResult_0(
+class _FakeWeather_0 extends _i1.SmartFake implements _i2.Weather {
+  _FakeWeather_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -33,24 +31,19 @@ class _FakeAppApiResult_0<T> extends _i1.SmartFake
         );
 }
 
-/// A class which mocks [WeatherRepository].
+/// A class which mocks [WeatherDataSource].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockWeatherRepository extends _i1.Mock implements _i3.WeatherRepository {
-  MockWeatherRepository() {
-    _i1.throwOnMissingStub(this);
-  }
-
+class MockWeatherDataSource extends _i1.Mock implements _i3.WeatherDataSource {
   @override
-  _i2.AppApiResult<_i4.Weather> getWeather(
-          {required _i5.WeatherRequest? request}) =>
+  _i2.Weather getWeather({required _i4.WeatherRequest? request}) =>
       (super.noSuchMethod(
         Invocation.method(
           #getWeather,
           [],
           {#request: request},
         ),
-        returnValue: _FakeAppApiResult_0<_i4.Weather>(
+        returnValue: _FakeWeather_0(
           this,
           Invocation.method(
             #getWeather,
@@ -58,5 +51,13 @@ class MockWeatherRepository extends _i1.Mock implements _i3.WeatherRepository {
             {#request: request},
           ),
         ),
-      ) as _i2.AppApiResult<_i4.Weather>);
+        returnValueForMissingStub: _FakeWeather_0(
+          this,
+          Invocation.method(
+            #getWeather,
+            [],
+            {#request: request},
+          ),
+        ),
+      ) as _i2.Weather);
 }
