+++ test/mocks/mock_weather_repository.dart
@@ -1,0 +1,8 @@
+// ignore_for_file: unused_import
+
+import 'package:flutter_training/data/repository/weather_repository.dart';
+import 'package:mockito/annotations.dart';
+import 'package:yumemi_weather/yumemi_weather.dart';
+
+@GenerateNiceMocks([MockSpec<WeatherRepository>()])
+import 'mock_weather_repository.mocks.dart';
