+++ test/mocks/mock_weather_repository.mocks.dart
@@ -1,13 +1,14 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in flutter_training/test/unit_test/data/repository/weather_repository_test.dart.
+// in flutter_training/test/mocks/mock_weather_repository.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:flutter_training/data/data_source/weather_data_source.dart'
-    as _i3;
-import 'package:flutter_training/data/models/weather/weather.dart' as _i2;
+import 'package:flutter_training/data/models/app_api_result.dart' as _i2;
+import 'package:flutter_training/data/models/weather/weather.dart' as _i4;
 import 'package:flutter_training/data/models/weather/weather_request.dart'
-    as _i4;
+    as _i5;
+import 'package:flutter_training/data/repository/weather_repository.dart'
+    as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -21,8 +22,9 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeWeather_0 extends _i1.SmartFake implements _i2.Weather {
-  _FakeWeather_0(
+class _FakeAppApiResult_0<T> extends _i1.SmartFake
+    implements _i2.AppApiResult<T> {
+  _FakeAppApiResult_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -31,23 +33,20 @@ class _FakeWeather_0 extends _i1.SmartFake implements _i2.Weather {
         );
 }
 
-/// A class which mocks [WeatherDataSource].
+/// A class which mocks [WeatherRepository].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockWeatherDataSource extends _i1.Mock implements _i3.WeatherDataSource {
-  MockWeatherDataSource() {
-    _i1.throwOnMissingStub(this);
-  }
-
+class MockWeatherRepository extends _i1.Mock implements _i3.WeatherRepository {
   @override
-  _i2.Weather getWeather({required _i4.WeatherRequest? request}) =>
+  _i2.AppApiResult<_i4.Weather> getWeather(
+          {required _i5.WeatherRequest? request}) =>
       (super.noSuchMethod(
         Invocation.method(
           #getWeather,
           [],
           {#request: request},
         ),
-        returnValue: _FakeWeather_0(
+        returnValue: _FakeAppApiResult_0<_i4.Weather>(
           this,
           Invocation.method(
             #getWeather,
@@ -55,5 +54,13 @@ class MockWeatherDataSource extends _i1.Mock implements _i3.WeatherDataSource {
             {#request: request},
           ),
         ),
-      ) as _i2.Weather);
+        returnValueForMissingStub: _FakeAppApiResult_0<_i4.Weather>(
+          this,
+          Invocation.method(
+            #getWeather,
+            [],
+            {#request: request},
+          ),
+        ),
+      ) as _i2.AppApiResult<_i4.Weather>);
 }
