+++ test/mocks/mock_yumemi_weather.dart
@@ -1,0 +1,7 @@
+// ignore_for_file: unused_import
+
+import 'package:mockito/annotations.dart';
+import 'package:yumemi_weather/yumemi_weather.dart';
+
+@GenerateNiceMocks([MockSpec<YumemiWeather>()])
+import 'mock_yumemi_weather.mocks.dart';
